refactor browser settings components & hooks structure
