+++ src/components/BrowserSettingsProvider/index.ts
@@ -1,0 +1,8 @@
+import { getSettings, setSettings } from "./settings";
+
+import useBrowserSettings from "./useBrowserSettings";
+import BrowserSettingsProvider from "./BrowserSettingsProvider";
+
+export default BrowserSettingsProvider;
+export { useBrowserSettings, getSettings, setSettings };
+export * from "./BrowserSettingsProvider";
